downgrade to 2.0.6

+++ Cache.php
@@ -71,7 +71,7 @@ class Cache extends \yii\caching\Cache
     /**
      * Initializes the redis Cache component.
      * This method will initialize the [[redis]] property to make sure it refers to a valid redis connection.
-     * @throws InvalidConfigException if [[redis]] is invalid.
+     * @throws \yii\base\InvalidConfigException if [[redis]] is invalid.
      */
     public function init()
     {
@@ -87,7 +87,7 @@ class Cache extends \yii\caching\Cache
      * may return false while exists returns true.
      * @param mixed $key a key identifying the cached value. This can be a simple string or
      * a complex data structure consisting of factors representing the key.
-     * @return boolean true if a value exists in cache, false if the value is not in the cache or expired.
+     * @return bool true if a value exists in cache, false if the value is not in the cache or expired.
      */
     public function exists($key)
     {
@@ -125,7 +125,7 @@ class Cache extends \yii\caching\Cache
         if ($expire == 0) {
             return (bool) $this->redis->executeCommand('SET', [$key, $value]);
         } else {
-            $expire = ($expire * 1000);
+            $expire = (int) ($expire * 1000);
 
             return (bool) $this->redis->executeCommand('SET', [$key, $value, 'PX', $expire]);
         }
@@ -146,7 +146,7 @@ class Cache extends \yii\caching\Cache
         if ($expire == 0) {
             $this->redis->executeCommand('MSET', $args);
         } else {
-            $expire = ($expire * 1000);
+            $expire = (int) ($expire * 1000);
             $this->redis->executeCommand('MULTI');
             $this->redis->executeCommand('MSET', $args);
             $index = [];
@@ -174,7 +174,7 @@ class Cache extends \yii\caching\Cache
         if ($expire == 0) {
             return (bool) $this->redis->executeCommand('SET', [$key, $value, 'NX']);
         } else {
-            $expire = ($expire * 1000);
+            $expire = (int) ($expire * 1000);
 
             return (bool) $this->redis->executeCommand('SET', [$key, $value, 'PX', $expire, 'NX']);
         }
